+++ server/datasources/cachedMongooseDataSource.ts
@@ -1,6 +1,7 @@
 import { Document, Types } from 'mongoose'
 import { Cache } from '../database/cache'
 import { DataSource } from 'apollo-datasource'
+import { ServerResponseInterface } from '../graphql/serverResponse';
 
 export default class CachedMongooseDataSource<DocumentType extends Document> extends DataSource {
   cache: Cache<DocumentType>;
@@ -16,6 +17,70 @@ export default class CachedMongooseDataSource<DocumentType extends Document> ext
 
   getAll(): Array<DocumentType> {
     return this.cache.getData()
+  }
+
+  async create(request: Document): Promise<ServerResponseInterface> {
+    const Request = this.cache.model
+    const newRequest = new Request(request)
+    const promise = await newRequest.save()
+      .then(res => {
+        return {
+          "success": true,
+          "message": this.cache.name + " successfully created",
+          "id": res._id
+        }
+      })
+      .catch(error => {
+        console.log(error)
+        return {
+          "success": false,
+          "message": error._message,
+          "id": null 
+        }
+      })
+    return promise
+  }
+
+  async update(request: Document): Promise<ServerResponseInterface> {
+    const Request = this.cache.model
+    const promise = await Request.findByIdAndUpdate(request._id, {...request, dateUpdated: Date.now()})
+      .then(res => {
+        return {
+          "success": true,
+          "message": this.cache.name + " successfully updated",
+          "id": request._id
+        }
+      })
+      .catch(error => {
+        console.log(error)
+        return {
+          "success": false,
+          "message": error._message,
+          "id": request._id
+        }
+      })
+    return promise
+  }
+  
+  async softDelete(id: Types.ObjectId) {
+    const Request = this.cache.model
+    const promise = Request.findByIdAndUpdate(id, {"deleted": true})
+      .then(res => {
+        return {
+          "success": true,
+          "message": this.cache.name + " successfully soft deleted",
+          "id": id
+        }
+      })
+      .catch(error => {
+        console.log(error)
+        return {
+          "success": false,
+          "message": error._message,
+          "id": id
+        }
+      })
+    return promise
   }
 }
